Add BackedProjectController.js && CreatedProjectController.js

+++ wip/Users/MaiCTP/DDL_CapstoneProject_Mai/DDL_CapstoneProject/Controllers/ApiControllers/ProjectApiController.cs
@@ -27,12 +27,12 @@ namespace DDL_CapstoneProject.Controllers.ApiControllers
 
             if (currentUser == null)
             {
-                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "Chưa đăng nhập!", Type = "UserNotFound" });
+               return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "Chưa đăng nhập!", Type = DDLConstants.HttpMessageType.NOT_AUTHEN });
             }
 
             if (!ModelState.IsValid)
             {
-                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = "Bad-Request" });
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
             }
 
             try
@@ -43,17 +43,19 @@ namespace DDL_CapstoneProject.Controllers.ApiControllers
             }
             catch (Exception)
             {
-                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = "Bad-Request" });
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
             }
             return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.SUCCESS, Message = "", Type = "", Data = id });
         }
 
         // PUT: api/ProjectApi/EditProject  
-        [ResponseType(typeof(ProjectDetailDTO))]
+        [ResponseType(typeof(ProjectEditDTO))]
         [HttpPost]
         public IHttpActionResult EditProjectBasic()
         {
             var httpRequest = HttpContext.Current.Request;
+
+           ProjectEditDTO updateProject = null;
 
             if (httpRequest.Form.Count <= 0)
             {
@@ -61,13 +63,28 @@ namespace DDL_CapstoneProject.Controllers.ApiControllers
             }
 
             var projectJson = httpRequest.Form["project"];
-            var project = new JavaScriptSerializer().Deserialize<ProjectDetailDTO>(projectJson);
+            var project = new JavaScriptSerializer().Deserialize<ProjectEditDTO>(projectJson);
 
+          if (project == null)
+            {
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
+            }
+			
             string imageName = "img_" + project.ProjectCode;
             var file = httpRequest.Files["file"];
             var uploadImageName = CommonUtils.UploadImage(file, imageName, DDLConstants.FileType.PROJECT);
 
-            var updateProject = ProjectRepository.Instance.EditProjectBasic(project, uploadImageName);
+  try
+            {
+                updateProject = ProjectRepository.Instance.EditProjectBasic(project, uploadImageName);
+            }
+            catch (Exception)
+            {
+
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
+            }
+
+            //var updateProject = ProjectRepository.Instance.EditProjectBasic(project, uploadImageName);
 
             return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.SUCCESS, Message = "", Type = "" });
         }
@@ -77,22 +94,68 @@ namespace DDL_CapstoneProject.Controllers.ApiControllers
         [HttpPut]
         public IHttpActionResult EditProjectStory(ProjectStoryDTO project)
         {
+			    ProjectStoryDTO updateProject = null;
+
             if (project == null)
             {
                 return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
             }
+ try
+            {
+                updateProject = ProjectRepository.Instance.EditProjectStory(project);
+            }
+            catch (Exception)
+            {
 
-            var updateProject = ProjectRepository.Instance.EditProjectStory(project);
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
+            }
+
+         //   var updateProject = ProjectRepository.Instance.EditProjectStory(project);
 
             return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.SUCCESS, Message = "", Type = "" });
         }
 
         // GET: api/ProjectApi/GetProject/:id
         [HttpGet]
-        [ResponseType(typeof(ProjectDetailDTO))]
+         [ResponseType(typeof(ProjectEditDTO))]
         public IHttpActionResult GetProject(int id)
         {
-            var project = ProjectRepository.Instance.GetProjectBasic(id);
+  ProjectEditDTO project = null;
+
+            var currentUser = getCurrentUser();
+
+            if (currentUser == null)
+            {
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "Chưa đăng nhập!", Type = DDLConstants.HttpMessageType.NOT_AUTHEN });
+            }
+
+            if (!ModelState.IsValid)
+            {
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
+            }
+
+            if (User.Identity == null || !User.Identity.IsAuthenticated)
+            {
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "Chưa đăng nhập", Type = DDLConstants.HttpMessageType.NOT_AUTHEN });
+            }
+
+            try
+            {
+                project = ProjectRepository.Instance.GetProjectBasic(id, currentUser.DDL_UserID);
+                if (project.ImageUrl != string.Empty)
+                {
+                    project.ImageUrl = DDLConstants.FileType.PROJECT + project.ImageUrl;
+                }
+            }
+            catch (ProjectNotFoundException)
+            {
+
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "Không tìm thấy dự án!", Type = DDLConstants.HttpMessageType.NOT_FOUND });
+            }
+            catch (Exception)
+            {
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
+            }
 
             return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.SUCCESS, Data = project });
         }
@@ -102,8 +165,16 @@ namespace DDL_CapstoneProject.Controllers.ApiControllers
         [ResponseType(typeof(ProjectStoryDTO))]
         public IHttpActionResult GetProjectStory(int id)
         {
-            var project = ProjectRepository.Instance.GetProjectStory(id);
+          ProjectStoryDTO project = null;
+            try
+            {
+                project = ProjectRepository.Instance.GetProjectStory(id);
+            }
+            catch (Exception)
+            {
 
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
+            }
             return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.SUCCESS, Data = project });
         }
 
@@ -112,7 +183,17 @@ namespace DDL_CapstoneProject.Controllers.ApiControllers
         [ResponseType(typeof(RewardPkgDTO))]
         public IHttpActionResult GetRewardPkg(int id)
         {
-            var rewardPkg = RewardPkgRepository.Instance.GetRewardPkg(id);
+           List<RewardPkgDTO> rewardPkg;
+
+            try
+            {
+                rewardPkg = RewardPkgRepository.Instance.GetRewardPkg(id);
+            }
+            catch (Exception)
+            {
+
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
+            }
 
             return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.SUCCESS, Data = rewardPkg });
         }
@@ -127,12 +208,12 @@ namespace DDL_CapstoneProject.Controllers.ApiControllers
 
             if (currentUser == null)
             {
-                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "Chưa đăng nhập!", Type = "UserNotFound" });
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "Chưa đăng nhập!", Type = DDLConstants.HttpMessageType.NOT_AUTHEN });
             }
 
             if (!ModelState.IsValid)
             {
-                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = "Bad-Request" });
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
             }
 
             try
@@ -141,7 +222,7 @@ namespace DDL_CapstoneProject.Controllers.ApiControllers
             }
             catch (Exception)
             {
-                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = "Bad-Request" });
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
             }
             return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.SUCCESS, Message = "", Type = "", Data = rewardPkg });
         }
@@ -151,12 +232,21 @@ namespace DDL_CapstoneProject.Controllers.ApiControllers
         [HttpPut]
         public IHttpActionResult EditRewardPkg(List<RewardPkgDTO> rewardPkg)
         {
+            bool result = false;
             if (rewardPkg == null)
             {
-                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = "Bad-Request" });
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
             }
 
-            var result = RewardPkgRepository.Instance.EditRewardPkg(rewardPkg);
+           try
+            {
+                result = RewardPkgRepository.Instance.EditRewardPkg(rewardPkg);
+            }
+            catch (Exception)
+            {
+
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
+            }
 
             return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.SUCCESS, Message = "", Type = "" });
         }
@@ -166,6 +256,13 @@ namespace DDL_CapstoneProject.Controllers.ApiControllers
         [HttpDelete]
         public IHttpActionResult DeleteRewardPkg(int id)
         {
+          var currentUser = getCurrentUser();
+
+            if (currentUser == null)
+            {
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "Chưa đăng nhập!", Type = DDLConstants.HttpMessageType.NOT_AUTHEN });
+            }
+
             try
             {
                 var result = RewardPkgRepository.Instance.DeleteRewardPkg(id);
@@ -173,7 +270,7 @@ namespace DDL_CapstoneProject.Controllers.ApiControllers
             }
             catch (Exception)
             {
-                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = "Bad-Request" });
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
             }
 
             return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.SUCCESS, Message = "", Type = "" });
@@ -184,7 +281,23 @@ namespace DDL_CapstoneProject.Controllers.ApiControllers
         [ResponseType(typeof(UpdateLogDTO))]
         public IHttpActionResult GetUpdateLog(int id)
         {
-            var updateLog = UpdateLogRepository.Instance.GetUpdateLog(id);
+           List<UpdateLogDTO> updateLog = null;
+            var currentUser = getCurrentUser();
+
+            if (currentUser == null)
+            {
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "Chưa đăng nhập!", Type = DDLConstants.HttpMessageType.NOT_AUTHEN });
+            }
+
+            try
+            {
+                updateLog = UpdateLogRepository.Instance.GetUpdateLog(id);
+            }
+            catch (Exception)
+            {
+
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
+            }
 
             return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.SUCCESS, Data = updateLog });
         }
@@ -195,25 +308,31 @@ namespace DDL_CapstoneProject.Controllers.ApiControllers
         public IHttpActionResult CreateUpdateLog(int id, UpdateLogDTO newUpdateLog)
         {
             var currentUser = getCurrentUser();
+			var updateLog = new UpdateLog();
 
             if (currentUser == null)
             {
-                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "Chưa đăng nhập!", Type = "UserNotFound" });
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "Chưa đăng nhập!", Type = DDLConstants.HttpMessageType.NOT_AUTHEN });
             }
 
             if (!ModelState.IsValid)
             {
-                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = "Bad-Request" });
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
             }
 
-            //try
-            //{
-            var updateLog = UpdateLogRepository.Instance.CreateUpdateLog(id, newUpdateLog);
-            //}
-            //catch (Exception)
-            //{
-            //    return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = "Bad-Request" });
-            //}
+           try
+            {
+
+
+                updateLog = UpdateLogRepository.Instance.CreateUpdateLog(id, newUpdateLog);
+            }
+
+            catch (Exception)
+            {
+
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
+            }
+         
             return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.SUCCESS, Message = "", Type = "", Data = updateLog });
         }
 
@@ -224,10 +343,18 @@ namespace DDL_CapstoneProject.Controllers.ApiControllers
         {
             if (updateLog == null)
             {
-                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = "Bad-Request" });
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
             }
 
-            var result = UpdateLogRepository.Instance.EditUpdateLog(updateLog);
+            try
+            {
+                var result = UpdateLogRepository.Instance.EditUpdateLog(updateLog);
+            }
+            catch (Exception)
+            {
+
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
+            }
 
             return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.SUCCESS, Message = "", Type = "" });
         }
@@ -248,25 +375,10 @@ namespace DDL_CapstoneProject.Controllers.ApiControllers
             }
             catch (Exception)
             {
-                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = "Bad-Request" });
-            }
-
-            return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.SUCCESS, Message = "", Type = "" });
-        }
-
-        // Post: Upload Image
-        [HttpPost]
-        public IHttpActionResult fileUpload(string ProjectCode)
-        {
-            var httpRequest = HttpContext.Current.Request;
-            if (httpRequest.Files.Count <= 0)
-            {
+ 
                 return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
             }
-            string imageName = "img_" + ProjectCode;
-            var file = httpRequest.Files[0];
-            var uploadImageName = CommonUtils.UploadImage(file, imageName, DDLConstants.FileType.PROJECT);
-
+ 
             return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.SUCCESS, Message = "", Type = "" });
         }
 
@@ -290,9 +402,33 @@ namespace DDL_CapstoneProject.Controllers.ApiControllers
         }
         public IHttpActionResult GetStatisticListForHome()
         {
-
             var listStatisticForHome = ProjectRepository.Instance.GetStatisticListForHome();
             return Ok(new HttpMessageDTO { Status = "success", Data = listStatisticForHome });
+        }
+
+        // PUT: api/ProjectApi/EditProjectStory  
+        [ResponseType(typeof(ProjectEditDTO))]
+        [HttpPut]
+        public IHttpActionResult SubmitProject(ProjectEditDTO project)
+        {
+            ProjectEditDTO updateProject = null;
+
+            if (project == null)
+            {
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
+            }
+
+            try
+            {
+                updateProject = ProjectRepository.Instance.SubmitProject(project);
+            }
+            catch (Exception)
+            {
+
+                return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.ERROR, Message = "", Type = DDLConstants.HttpMessageType.BAD_REQUEST });
+            }
+
+            return Ok(new HttpMessageDTO { Status = DDLConstants.HttpMessageType.SUCCESS, Message = "", Type = "" });
         }
 
         //Trungvn
